Add andymccurdy/redis-py@8ea26c4f5c82a3fde5dd2fb209d0c8644714626d Migrating commands to a mixin (andymccurdy/redis-py#1534)
* Fixes #1136
* Added typing.py to store type annotations
