+++ aioredis/compat.py
@@ -2,4 +2,8 @@
 try:
     from typing import Literal, Protocol, TypedDict  # lgtm [py/unused-import]
 except ImportError:
-    from typing_extensions import Literal, Protocol, TypedDict  # lgtm [py/unused-import]
+    from typing_extensions import (  # lgtm [py/unused-import]
+        Literal,
+        Protocol,
+        TypedDict,
+    )
